refactor: divide Judge class
